requirements.txt
# Core environment
google-adk
python-dotenv==1.0.0

# LangChain
langchain
langchain-community
langchain-openai

# OpenAI SDK
openai==1.35.13

# LlamaIndex and agents (compatible versions)


# Embeddings support
sentence-transformers==2.5.1

# Search
google-search-results
serpapi==0.1.0

# Google dependencies
google-api-core==2.17.1
pydantic-core==2.16.3
